% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2020_Assignment_3.R
\name{dijkstra}
\alias{dijkstra}
\title{Find the shortest path Between vertices using Dijkstra's algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Data frame consisting vertices and distance between them}

\item{init_node}{Initial Node/vertex from which the distance is to be calculated}
}
\value{
Shortest distance from the initial node/Vertex to all other vertices
}
\description{
Find the shortest path Between vertices using Dijkstra's algorithm
}
\details{
{This function finds the shortest distance from the input 
vertex to all other vertices.
   
For additional information click on the below URL

\url{https://en.wikipedia.org/wiki/Graph_(mathematics)}
}
}
\examples{
\dontrun{ 
wiki_graph <- data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6), 
                         v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
                         w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))
dijkstra(wiki_graph,1)
0 7 9 20 20 11}

}
