% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2020_Assignment_3.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm to find the greatest common divisor of two numbers}
\usage{
euclidean(int1, int2)
}
\arguments{
\item{int1}{Input 1 - Scalar Integer value}

\item{int2}{Input 2 - Scalar Integer value}
}
\value{
Returns the greatest common divisor of \code{int1} and \code{int2}
}
\description{
Euclidean algorithm to find the greatest common divisor of two numbers
}
\details{
{This function finds greatest common divisor of the given 2 numbers. If inputs are not numeric scalar,
the program returns error

For additional information click on the below URL
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
}
\examples{
\dontrun{
euclidean(123612, 13892347912)
4}

If the user inputs decimal values, the program rounds it to the integer.
}
