% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate a vector}
\usage{
rotate(x, k)
}
\arguments{
\item{x}{a vector}

\item{k}{an integer, the position to rotate}
}
\value{
a vector
}
\description{
Rotate a vector
}
\examples{
rotate(1:7, 3)

}
