% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transformations}
\alias{transformations}
\alias{scan}
\alias{rotate}
\title{Transformations}
\usage{
scan(x, f)

rotate(x, i)
}
\arguments{
\item{x}{a vector}

\item{f}{a function}

\item{i}{indices}
}
\value{
a vector
}
\description{
Transformations
}
\examples{
scan(1:5, "+")
rotate(1:5, 2)
}
