% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{
The data frame contains 18 observations of 3 variables that
describe the graph's edges and their weights:

v1: The starting vector of the edge

v2: The destination vector of the edge

w: The weight of the edge

This example graph is based on the one shown at the top of this wikipedia
article: \url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A data frame that represents an example graph to use with the dijkstra()
function.
}
\keyword{datasets}
